+++ example/src/demo/NumberFormat.example.js
@@ -1,22 +1,20 @@
 import React from 'react';
 import { Utils } from 'bng-react-lib';
-import { FormLabel, Container } from '../common/Form';
+import { FormLabel, Container, Title } from '../common/Form';
 
 let { makeNumberFormat, LOCALES, numberDecimal } = Utils; // LOCALES.enUS, LOCALES.deDE
 
 const NumberFormatExample = () => {
-  const [searchKey, setSearchKey] = React.useState('1001');
+  const [value, setValue] = React.useState('10');
   const [minimumFractionDigits, setMinimumFractionDigits] = React.useState(2);
   const [maximumFractionDigits, setMaximumFractionDigits] = React.useState('');
 
   const numberFormatEnUS = makeNumberFormat(minimumFractionDigits, maximumFractionDigits, LOCALES.enUS);
   const numberFormatDeDE = makeNumberFormat(minimumFractionDigits, maximumFractionDigits, LOCALES.deDE);
 
-  const handleSearch = (e) => {
-    setSearchKey(e.target.value);
-  }
-
   return <div>
+    <Title>Note: Use makeNumberFormat, LOCALES, numberDecimal in Utils</Title>
+
     <Container inline>
       <FormLabel label='minimumFractionDigits:'>
         <input type='number' value={minimumFractionDigits} onChange={e => setMinimumFractionDigits(e.target.value)} />
@@ -25,20 +23,22 @@ const NumberFormatExample = () => {
         <input type='number' value={maximumFractionDigits} onChange={e => setMaximumFractionDigits(e.target.value)} />
       </FormLabel>
     </Container>
+
     <hr />
+
     <FormLabel label='Input number:' >
-      <input value={searchKey} onChange={handleSearch} />
+      <input type='number' value={value} onChange={e => setValue(e.target.value)} />
     </FormLabel>
 
     <FormLabel label='makeNumberFormat by en-US' inline>
-      {numberFormatEnUS(searchKey)}
+      {numberFormatEnUS(value)}
     </FormLabel>
     <FormLabel label='makeNumberFormat by de-DE' inline>
-      {numberFormatDeDE(searchKey)}
+      {numberFormatDeDE(value)}
     </FormLabel>
 
     <FormLabel label='numberDecimal' inline>
-      {numberDecimal(searchKey, minimumFractionDigits)}
+      {numberDecimal(value, minimumFractionDigits)}
     </FormLabel>
   </div>
 }