+++ src/components/index.js
@@ -4,3 +4,4 @@ export { VirtualizedTable } from './VirtualizedTable';
 export { MultiGridTable } from './MultiGridTable';
 export { ReactSortableTree } from './ReactSortableTree';
 export { VirtualizedSelect } from './VirtualizedSelect';
+export { EllipsisTooltip } from './EllipsisTooltip';